Generic vite hosting for all viewer test apps
