+++ packages/tools/viewer-alpha/package.json
@@ -5,9 +5,9 @@
     "sideEffects": false,
     "description": "A viewer using BabylonJS to display 3D elements natively",
     "scripts": {
-        "start:web": "npm run serve:web -- --open",
-        "serve:web": "vite --config test/apps/web/vite.config.js",
+        "start:web": "npm run serve -- --open test/apps/web/index.html",
         "preserve": "npm run build",
+        "serve": "vite",
         "build": "npm run compile",
         "clean": "rimraf dist && rimraf *.tsbuildinfo",
         "compile": "tsc -b tsconfig.build.json",
@@ -19,6 +19,7 @@
     "devDependencies": {
         "@dev/core": "1.0.0",
         "@dev/loaders": "1.0.0",
+        "chalk": "^5.3.0",
         "source-map-loader": "^4.0.0",
         "ts-loader": "^9.2.6",
         "vite": "^5.3.3"
