+++ packages/tools/viewer-alpha/vite.config.mjs
@@ -1,0 +1,22 @@
+import { defineConfig, loadEnv } from "vite";
+import chalk from "chalk";
+
+export default defineConfig(({ mode }) => {
+    const env = loadEnv(mode, process.cwd());
+    const source = env.source ?? "dev";
+
+    const port = env.VIEWER_PORT ?? 1342;
+    console.log(`${chalk.bold(`Web Test App`)}: ${chalk.cyan(`http://localhost:${port}/test/apps/web/index.html`)}`);
+
+    return {
+        server: {
+            port,
+        },
+        resolve: {
+            alias: {
+                core: `@${source}/core/dist`,
+                loaders: `@${source}/loaders/dist`,
+            },
+        },
+    };
+});
